pyproject.toml
[project]
name = "civics"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "ollama>=0.6.1",
]

[dependency-groups]
dev = [
    "pdf2image>=1.17.0",
    "pillow>=12.2.0",
    "pytesseract>=0.3.13",
]
